Time: 29 ms (49.82%), Space: 48.5 MB (17.85%) - LeetHub

+++ 1283-find-the-smallest-divisor-given-a-threshold/1283-find-the-smallest-divisor-given-a-threshold.java
@@ -24,9 +24,9 @@ class Solution {
 
     }
 
-    public static double check(int[] arr,int mid)
+    public static int check(int[] arr,int mid)
     {
-        double sum=0;
+        int sum=0;
         for(int i=0;i<arr.length;i++)
         {
             sum+=Math.ceil((double)arr[i]/mid);
